Update patch set 11
Patch Set 11:

Eric, so I have some questions about that patch... First, it's failing with some error that was not happening before. See https://jenkins.akraino.org/job/ci-management-jjb-verify/117/console, 10:16:35 IOError: [Errno 21] Is a directory: 'archives/job-configs/kni' . I don't know why it happens, and how to fix it, can you help?
Second, i'd like that this job is executed on each code change, and also when it's submitted. But i'd only want to publish the generated binary when it's submitted, how can i achieve it? What should be the better place to store it? Thanks

Patch-set: 11