Update patch set 14
Patch Set 14:

Eric, i had to remove the initial view , otherwise the patch was failing (see previous runs). If you have any idea on how to fix it, please tell me. 
Second, i'd like that this job is executed on each code change, and also when it's submitted. But i'd only want to publish the generated binary when it's submitted, how can i achieve it? What should be the better place to store it? Thanks

Patch-set: 14