Update patch set 15
Patch Set 15:

Yolanda, I uploaded a changeset with some fixes. However, the job is still failing due to a bad path. But that's specific to this job, so I'm passing it back to you. You can see the output here:

https://jenkins.akraino.org/sandbox/view/kni/job/kni-installer-build-installer/4/console

In terms of uploading only after merge, I would suggest making a "common" job template, and then creating one sub-job called "verify" and one called "merge", where only the latter adds a builder to push the artifact. For an example of this sort of inheritance, take a look at this file from global-jjb, particularly how lf_jenkins_cfg_merge is setup, and then gerrit-jenkins-cfg-merge and github-jenkins-cfg-merge inherit that setup while adding gerrit/github-specific config to it.

https://github.com/lfit/releng-global-jjb/blob/d7ab98245ca71895ab0fce4cf23c41a888a1e7f3/jjb/lf-ci-jobs.yaml#L267

Patch-set: 15