Update patch set 2
Patch Set 2: Code-Review-1

I'd recommend just one more change to the commit message: in the description, add some detail as to _why_ the change is being made. Basically, sum up the reasons that are listed in the ticket (that the "project" variable allows slashes, and the job names can't have slashes in them).

Patch-set: 2
Label: Code-Review=-1