[ta] Move image build job to a bigger build minion
TA image build job runs out of memory when the executor
has 8 GB memory. It has been tested to work with 16 GB.

The minion label centos7-builder-4c-16g is not available in
Akraino sandbox Jenkins but if it is available in production
Jenkins this change should fix the issue. Otherwise a flavor
with 16 GB RAM should be created in Akraino Jenkins.

Change-Id: Idc7a5855f47c6c0e539677a71dd989627985c14d
Signed-off-by: Ossi Koivistoinen <ossi.koivistoinen@nokia.com>