Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-2
> > 
> > Our log paths are designed in a specific way. Trying to work around the system will end up giving you bad data in returned information from Jenkins and in Gerrit.
> 
> I agree and I want to avoid that. Btw, Jenkins is already returning bad data to Jenkins, the comments added by Jenkins point to logs.akraino.org/releng, while they should point to logs.akraino.org/production or logs.akraino.org/sandbox, but that's out of the scope for this change.

That is easily fixed and is a configuration error on our part, in fact I just fixed what Jenkins is posting back to Gerrit. This should have been raised as an RT when folks noticed the problem!

> According to my information, for the first Akraino release, we are required to push logs to logs.akraino.org/arm/job/... instead of the default logs.akraino.org/production/<slave-name>/...
> The wiki page [1] suggests using an explicit `lftools deploy` call, but as far as I understand, that will also bypass the regular scheduled cleanup performed every X months.

I would really like to know the reasoning behind mandating a log path that is not part of our environmental configuration. Frankly, as the RE lead I will _not_ let this change into the environment. It's completely silly and wrong.

As for the log cleanups. I fixed an issue with how our cleanups are performed last week. All logs that hit the system will now be properly cleaned up at the expected time.

> So, do you have any suggestion on how we can push to that log path without causing additional issues?

Yes, you don't.

Patch-set: 1