Update patch set 9
Patch Set 9:

Is there a way to avoid moving out of the workspace? If not, the script needs to include cleanup. I was thinking that using HOME wasn't a big deal due to the build nodes being one-time use, but this is going to build on your static node. It would be _extremely_ bad practice to build in a static location and not clean up afterwards.

Patch-set: 9
Label: Verified=0