Update patch set 2
Patch Set 2:

This doesn't change the node label, so it is still looking for any node labeled "aarch64_dev". Will all jobs run properly on those nodes?

Also, in the job iec-deploy-fuel-baremetal-centos7-daily-master (I'm sure this applies to many other jobs as well) I see a hardcoded reference to Ubuntu 16.04. In the second build script, the following line appears:
export ADAPTER_OS_PATTERN='(?i)ubuntu-16.04.*arm.*'
Is this going to be a problem? Should further changes to this script be made?

Patch-set: 2