Update patch set 2
Patch Set 2:

(1 comment)

Sorry for my incomplete understanding of how all the macros interact, but when I look in the Jenkins GUI for the ta-ci-build job I see the string "@daily" in the box for "Build Triggers". I would have expected to see something in this change where "@daily" or something equivalent was removed from the YAML and replaced with something that checks the "Build when another project is promoted" checkbox.

Can you confirm that the expected result of this change is that instead of building on a fixed time schedule, this change will build a new ISO whenever a change is merged into any of the TA Gerrit repos.

Patch-set: 2
Reviewer: Gerrit User 43 <43@4cada01b-0314-4fe0-84a1-f9ddc89863f9>
Label: Code-Review=0