Update patch set 1
Patch Set 1:

What we need on this patch, is to be triggered just by a comment . We don't want the job to be triggered when a patch is created, etc... so that's the reason for removing those triggers. How would you do it instead?

Patch-set: 1
Reviewer: Gerrit User 140 <140@4cada01b-0314-4fe0-84a1-f9ddc89863f9>
Label: Code-Review=0