Update patch set 1
Patch Set 1:

I'm recommending that, in addition to the changes already made, the variable is renamed from "gerrit_verify_triggers" to "gerrit_build_triggers", or something else that better describes its function after the change. It is no longer a set of "verify triggers", so it should not retain that name.

Patch-set: 1