Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Is it necessary to change in every merge job? 
> Cant we have a template or a macro defined in jjb/akraino-templates/akraino-ta-common-macros.yaml within ta-builder-job-merge to call the ta/build-tools/build_images.sh script or the job ta-ci-build itself?

The merge jobs are design for a diff purpose and so does the ta-ci-build jobs, imho both should remain separate functional jobs/units and should be trigger at the end of the merge job. It's generally not a best practice to combine both together, and not done that with other LF projects. Assuming if build for the merge job fails, this would require the entire merge to start again, when both a clubbed together.

Patch-set: 1