Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Couldn't this just be added to the job template, as a conditional when project-type==merge?

yes, we could. However, do we want to trigger ci-build job for every verify job? No.

Adding the section directly to the template is hardcoding the object/jobs within the template, making less flexible for having diff project call diff jobs or add a job list for a few of the projects.

Patch-set: 1