Update patch set 2
Patch Set 2:

Currently testing in sandbox, I'm still waiting on change #1215 to be merged, but I managed to test using a personal repo.
There are a bunch of issues which seem to be specific to both virtual pods and arm64 pods. I will detail the various issues in Jira, but hopefully we could merge this in the meantime.

Patch-set: 2
Reviewer: Gerrit User 147 <147@4cada01b-0314-4fe0-84a1-f9ddc89863f9>
Label: Verified=0