Update patch set 3
Patch Set 3:

> Hi Eric,
 > 
 > With the current logic we are able to identify duplicates & only
 > upload the changed RPMs that are not present in Nexus but there is
 > a failure 403 forbidden error from Nexus. Can you please look into
 > it?
 > Link - https://jenkins.akraino.org/sandbox/view/All/job/ta-caas-danm-master-merge/3/console
 > 
 > Regards,
 > Amin

That's expected behavior on the sandbox. The sandbox does not have credentials to push to repos, because we don't want test jobs to push. As long as only the artifacts that you want to be pushed are showing up in the lftools output, then it's a passed test.

Patch-set: 3