Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > This is a bit trickier to get right, since all our (static) aarch64
>  > build nodes are currently running Ubuntu, which is not very
>  > friendly to building RPMs ...
>  > I will push another updated draft and continue debugging this in
>  > the sandbox; will ping back when all my tests are passing.
> 
> Do we need a new builder node type added to run ARM builds on CentOS or are you going to try to get RPM builds working on Ubuntu?
> 
> There is a project that has been around for a while that I've never used but which claims to handle package format translations. https://wiki.debian.org/Alien I don't know if that would be helpful.

Unfortunately `alien` does not help in this case, the issues I had were with getting the right version of mock to work on Ubuntu, since Xenial ships with a version that is too old + Ubuntu lacks the `rpmdevtools` package.
But I'll get `mock` to work on Ubuntu, no worries. It's easier than setting up a different build machine just for this.

Patch-set: 2