Update patch set 2
Patch Set 2:

> Why specify verify-jobs in the file name, or "-tox-verify" in the
 > name field? This only seems to limit expanding the file/jjb
 > project.

This specific jjb project will be used only for verifying the portal-onapsdk project based on a tox.ini file. I used the same approach as with the one used for the validation project

https://gerrit.akraino.org/r/gitweb?p=ci-management.git;a=blob;f=jjb/validation/validation-verify-jobs.yaml;h=2d72cc85b6b1ba6de092a6c37366a1e40fa5bc81;hb=refs/heads/master

The jjb project that will build and upload the war and jar files and the jjb project that will create the docker images will be different from the verification jjb. I will upload them in the future.

Patch-set: 2
Label: Verified=0