Update patch set 3
Patch Set 3:

The integration part with the UI is missing. Also, the upload of results into Nexus is missing as well. I believe that these two parts should be included in this patch.

So, maybe a redesign is needed. I propose to place under the ci-management/jjb/validation the master Jenkins job that will trigger the slave Jenkins jobs. The latter ones will be available in LF Jenkins only for labs that are integrated with this Jenkins. The labs that are not integrated with the aforementioned Jenkins, won't support such jobs under public scope. The administrators of these labs will be responsible for downloading the slave Jenkins jobs from LF jenkins and integrating them with their own private Jenkins.

Maybe we can discuss further during the community call (October 9, 2019)

Patch-set: 3