Update patch set 3
Patch Set 3:

> > Patch Set 3:
 > >
 > > The integration part with the UI is missing. Also, the upload of
 > results into Nexus is missing as well. I believe that these two
 > parts should be included in this patch.
 > >
 > > So, maybe a redesign is needed. I propose to place under the
 > ci-management/jjb/validation the master Jenkins job that will
 > trigger the slave Jenkins jobs. The latter ones will be available
 > in LF Jenkins only for labs that are integrated with this Jenkins.
 > The labs that are not integrated with the aforementioned Jenkins,
 > won't support such jobs under public scope. The administrators of
 > these labs will be responsible for downloading the slave Jenkins
 > jobs from LF jenkins and integrating them with their own private
 > Jenkins.
 > >
 > > Maybe we can discuss further during the community call (October
 > 9, 2019)
 > 
 > Thanks for the comment Ioakeim. I agree that we need to add the
 > part that pushes logs to nexus, but let's do that in a different
 > patch.
 > 
 > Regarding the integration with UI, my understanding was that UI
 > reads from the Jenkins jobs and exposes the capability to manually
 > trigger the jobs (a.k.a the UI has the rights to trigger a job that
 > already exists in Jenkins master). What kind of notification do you
 > need from Jenkins to UI? In my opinion, the job does not need to
 > know of the UI existence.

Hi Cristina,

before continuing this discussion, it should be noted that for the release of November 2019, we will not support UI mode 1, that is UI submits a blueprint for validation and it retrieves results from Nexus. We will only support UI mode 2, that is UI retrieves results from Nexus that every blueprint owner has independently pushed. However, it should be good if we could design the jjbs in such a way that UI mode 1 can be easily supported in the future.

So, the complete loop for mode 1 is: UI submits a blueprint for validation to the LF master Jenkins job. Then, this job triggers the appropriate slave Jenkins based on the lab input parameter. The slave Jenkins job can be in LF Jenkins if the corresponding lab is integrated with LF Jenkins, or it can be located in a private Jenkins instance if the lab is not integrated with LF Jenkins. On the contrary, for UI mode 2, the corresponding slave Jenkins is manually triggered by the blueprint owner.

Based on this description, this patch concerns the slave Jenkins jobs. In mode 1, these Jenkins jobs must be enabled to inform the UI when they upload results to Nexus. That's why we need integration with the UI. But as I have already mentioned, let us forget about mode 1 for now. I just mentioned it in order to make sure that the current patch can be future proof. Therefore, let us also forget about Jenkins master job. I think we should only mention that this patch is for slave Jenkins bluval validation test executions.

Also, I, Naga, Deepak and Indu have decided to propose to use this patch for slave Jenkins jobs in LF Jenkins. However, we should use the Jenkins job described here https://gerrit.akraino.org/r/#/c/validation/+/1647/ in cases where the blueprint owner's lab is not integrated with LF Jenkins.

I think we should discuss further during tomorrow's community call.

Patch-set: 3