Update patch set 9
Patch Set 9: Code-Review-1

Please put icn-template into the akraino-templates directory, and change the name to match the other files in that directory.
This should also be tested on the sandbox. For additional information, see the below guide. I have already added you to the access group for the Akraino sandbox.
https://docs.releng.linuxfoundation.org/en/latest/jenkins-sandbox.html

Patch-set: 9
Label: Code-Review=-1