Update patch set 2
Patch Set 2:

(1 comment)

This isn't really a comment on this particular change, but I don't like the fact that this job holds its own templates. Templates should be stored in "jjb/akraino-templates". I don't want to do it without warning, but I can tell you now that the next time you need to make any changes to this file, I'm going to ask that the template be moved into an entirely separate file before I merge any more changes. It may be a good idea to note this as internal tech debt, and handle it before the next time you need to change the template.

It's also worth noting that this really isn't a difficult change to make. Since jjb pretty much flattens everything at build time, you just need to cut all the template code out and put it into a separate file in akraino-templates.

Patch-set: 2
Label: Verified=0