Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> The merge job ran successfully when it was merged. Is there an issue?

tl;dr Looks like Jenkins does not trigger the EXIT trap added in this change.

Note that jjb-merge is completely irrelevant in this case, since the run_bluval.sh script is NOT embedded in the jenkins job, but instead invoked with something like `./run_bluval.sh ...`.
So as long as the change got into the git repository, the latest version will always be picked up automatically, without even updating the job definition via jjb-merge.

As for why the trap is not triggered, it might have something to do with the way Jenkins sends kill signals to its child processes, e.g. see [1], although that one describes the `job abort` use-case and not a regular run.

We could just take the easy way out here and do something simpler without traps, but there's a catch (see below):

  set +e
  # here's a catch: if someone aborts the job after it created the things that need cleanup
  # the finish code won't run
  python3 validation/bluval/blucon.py $options "$blueprint_name"
  exit_code=$?
  set -e
  finish
  exit $exit_code

To try to workaround that, we could keep the trap, but make it trigger on TERM.
However, this seems to be problematic too, see [2].

Would it be possible to just do `umask 000` inside the container that creates root-owned files?
Or bind-mount a temporary directory (`mktemp -d`) inside that container, then move the resulting files from /tmp/tmp.xxxxx/ to $WORKSPACE after container exit?

[1] https://gist.github.com/datagrok/dfe9604cb907523f4a2f
[2] https://issues.jenkins-ci.org/browse/JENKINS-17116

Patch-set: 2