Update patch set 1
Abandoned

This change does not need to be made. There was a problem with the Jenkins plugin for the SonarQube scanner, which has now been fixed. However, please take a look at the logs, as it seems there are some errors in the job. It also seems to be hanging at logs-deploy.sh.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon