Update patch set 4
Patch Set 4:

> Patch Set 3: Code-Review-1
> 
> You should just use the lf-infra-publish publisher. Rewriting this into a shell script is a very convoluted way to accomplish this task.
> Your job templates should also be moved to jjb/akraino-templates (which already has akraino-validation-{macros,templates}.yaml files).

We have distributed labs (Jenkins slaves and clusters), out of which some use their own Jenkins master, some use the LF one. For the labs that use the LF Jenkins master, when using lf-infra-publish the results will be pushed by default to nexus at production/vex-yul-akraino-jenkins-prod-1/

But the validation project needs the logs in a specific location (lab silo) in order to retrieve them further in the UI and process the test results. So this is why we push the validation logs with our customized script.

Patch-set: 4
Label: Verified=0