Update patch set 2
Patch Set 2: Code-Review-1

I need more time to think about this. I understand the problem you're trying to solve, but I'm not sure I like moving from simple integer build IDs to the date and timestamp format.

Also, if we're changing the destination of build products, we should probably eliminate the "release-1" part of the path at the same time. We're now past release 2 and heading for release 3, but the CI job is still pushing builds to a directory with "release-1" in the name.

Patch-set: 2
Label: Code-Review=-1