[WIP] Enable installation of TA in CI
For now, only UNH Ampere1 POD is added.

For each REC POD, the CD job is tied to a specific jenkins slave
that has access to the POD's lab.
The jenkins slave will store required environment in:
  /data/rec_pod/env.rc:
    HOST_IP=<JUMPHOST-IP>
    POD_NAME=<POD-NAME>
    ADMIN_PASSWD=<INITIAL-ADMIN-PASSWD>

For now, user_config.yaml is stored on the slave at:
  file:///data/rec_pod/user_config.yaml

Slave package requirements:
  - docker
  - jq

Co-authored-by: Cristina Pauna <cristina.pauna@enea.com>
Signed-off-by: Alexandru Antone <Alexandru.Antone@enea.com>
Change-Id: Iece4330c3242385c33d8a1daa092f78ce5d46cc8