Don't run lf-infra-publishers on deploy jobs
The deploy jobs are run on dedicated hosts that don't have the
pre-requisites needed by the lf-infra-publishers (eg, facter,
xmlstarlet, click).  We could just install the pre-reqs, but some
of the publishers (metadata collection for example) make the
assumption that the job is running in one of the cloud environments
and would need updates to work on the baremetal hosts used by these
jobs.

The post-build tasks don't seem important for these jobs, so just
skip them entirely.

Signed-off-by: James E. Blair <jeblair@redhat.com>
Change-Id: Ia7603a70fbe1993d154d2bbafa50026be1627599