Don't use sudo in baremetal installer job
This job is currently failing because it's attempting to access
files in ~root.

The build step prior to the one changed here installs the pull
secret and ssh key into the "jenkins" user's home directory.
The kni_deploy_baremetal.sh script expects to use the previously
prepared git repos also in the jenkins home directory.  It includes
a "sudo" command inside kni_deploy_baremetal.sh to run
virsh-cleanup.sh.  All of these suggest that this script should be
run as the jenkins user and it's expected to internally sudo when
necessary, so this change removes the sudo in the JJB definition.

Signed-off-by: James E. Blair <jeblair@redhat.com>
Change-Id: If62e1a9e7291fbbd9929c834552d3b9ee5932a50