Replace python-decoratortools per error message
According to the error message during the ARM64 verify job,
python-decoratortools is not available but the package
python-peak.util.decorators replaces it.

See
https://jenkins.akraino.org/job/ta-infra-ansible-arm64-master-verify/47/console
for details of the error.

Additionally, there seems to be an error due to a conflict between the
Ubuntu 18.04 of the builder VM vs the CentOS packages we're trying to
build. It seems the conflict didn't occur on the older Ubuntu 16.04
builder VMs that we used to use. Since the builder VMs get thrown away
and we don't really care about them after we finish building, we'll
force overwrite the Ubuntu version of '/usr/bin/annotate-output' with
the CentOS copy that is coming from the CentOS RPM that we care about.

Also adding a debugging step of listing Python packages due to an error
being reported about lxml.html import failing.

Signed-off-by: Paul Carver <pcarver@att.com>
Change-Id: Ib4e446190f718ef326a02087dc88fbfd8bdd293e