Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Has this been tested to ensure that it works as expected?

I'm not really sure how to run the Jenkins job outside of Jenkins.

What I have done is verified that running "apt install python-decoratortools" on my own Ubuntu VM fails with the same error message that I see in the Jenkins log and running "apt install python-peak.util.decorators" successfully installs the package.

I don't actually know what that package does or why it's required for the ARM64 build but not the AMD64 build, so to actually test whether the substitution works we need to build a new ARM64 image and try to deploy it. But if the substitution doesn't work, it won't do any good to leave apt install as-is because it simply fails because the package simply isn't available to be installed anymore. I assume that the upstream Ubuntu repository has removed it.

Patch-set: 1