Update patch set 2
Patch Set 2:

Eric,

Is there any way to run this job on a Jenkins build node with snd-ubuntu16.04-docker-arm64-8c-8g-1?

My understanding is that this job was originally developed on Ubuntu 16.04 and that the LF moved the build nodes to 18.04 which broke this job. I would like to verify if it will still succeed on 16.04.

I don't think it was a change to REC/TA that broke this job.

I've made two changes in an attempt to fix this job and now I'm running into a third distinct error which is just dumb, Python is failing to import the module lxml.html, but only on Arm, not on AMD64, and we don't think we've made any changes that should be responsible for removing a Python module.

Patch-set: 2