Move install of "rename" into ARM conditional
In order to fix the ARM build it was necessary to install the "rename"
command, but that was done via apt which does not exist on CentOS (the
ARM build is done on Ubuntu, the AMD build on CentOS) and the "rename"
command is not required for the AMD build. This change fixes that
mistake.

Signed-off-by: Paul Carver <pcarver@att.com>
Change-Id: I49755198a5774a611ae5977d7f17645461d08302