Chore: Upgrade global-jjb to v0.77.0
This version brings in the latest version of JJB, which is required
for upgrading to the latest versions of some plugins.

For compatibility with the latest version of global-jjb, this update
includes the creation of the ansible.cfg managed file, and updates
the default version of JJB used by CI jobs.

Change-Id: I9af9386d8a171d9487a130edd7577dc1f5088040
Signed-off-by: Eric Ball <eball@linuxfoundation.org>