Revert "CI: Rename global-settings to mvn-glob..."
This reverts commit b2b26d4b275f686897b9cc88649cd4f68370ea6a.

Though some jobs have mvn-global-settings listed as the global
settings file, renaming to "global-settings" is a standard that will
carry forward to all projects for Github Actions.

Since we are currently transitioning to GHA, this shouldn't be
renamed. If JJB jobs still need to use this file, they should be
modified to use "global-settings."

Change-Id: Ib1190de829939d11b022361ebf72868e7152c3cc
Issue-ID: LF-Jira RELENG-4799
Signed-off-by: Eric Ball <eball@linuxfoundation.org>